Extract Pact.Analyze.Model from Check
